feat: 서치바 클릭 시 여행지 추천 화면 보이도록 구현 (ios-h/airbnb#21)
